using some traits
