+++ src/PHPBot/Command/Keyboard/Linux/SendKeys.php
@@ -38,16 +38,5 @@ class SendKeys implements CommandInterface
         return $cmd;
     }
 
-    public function start()
-    {
-        $command = $this->getCommand($this->keys);
-        $deferred = new Deferred();
-        $command->start($this->loop, 0.001);
-
-        $command->on('exit', function () use ($deferred) {
-            $deferred->resolve();
-        });
-
-        return $deferred->promise();
-    }
+    use \PHPBot\Command\Keyboard\Traits\MultipleButtonsStartTrait;
 }
