+++ src/PHPBot/Command/Keyboard/Linux/Type.php
@@ -5,7 +5,6 @@ namespace PHPBot\Command\Keyboard\Linux;
 use PHPBot\Command\CommandInterface;
 
 use React\EventLoop\LoopInterface;
-use React\Promise\Deferred;
 use React\ChildProcess\Process as ChildProcess;
 
 class Type implements CommandInterface
@@ -28,16 +27,5 @@ class Type implements CommandInterface
         return $cmd;
     }
 
-    public function start($delay = 12)
-    {
-        $command = $this->getCommand($this->text, $delay);
-        $deferred = new Deferred();
-        $command->start($this->loop, 0.001);
-
-        $command->on('exit', function () use ($deferred) {
-            $deferred->resolve();
-        });
-
-        return $deferred->promise();
-    }
+    use \PHPBot\Command\Keyboard\Traits\TypeTextTrait;
 }
